github action runners can be self-hosted. i am running one inside my server.
turns out, the github action cannot just assume that it's the only one who's using
the server's infrastructure.

this should help w/ the ssh stuff.
fyi, i am not sure 100% if this fixes it:
i tried running the commands here and it helped,
but not sure if it i'll be the case when the action does so, but should be good.

for _not_ overriding the .ssh/config,
for _not_ spamming the .ssh/known_hosts file,

i'll consider creating separate PRs.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>